[language][functional tests] move the ir and move-lang harnesses into the library crate
This moves the ir and move-lang functional test harnesses into the
library crate. This is to allow them to be used by additional call
sites, making it easier for us to move tests around.
