+++ language/ir-testsuite/Cargo.toml
@@ -10,18 +10,10 @@ publish = false
 edition = "2018"
 
 [dev-dependencies]
-anyhow = "1.0.38"
-bytecode-verifier = { path = "../bytecode-verifier" }
-datatest-stable = "0.1.1"
-functional-tests = { path = "../testing-infra/functional-tests" }
-ir-to-bytecode = { path = "../compiler/ir-to-bytecode" }
-diem-types = { path = "../../types" }
 diem-workspace-hack = { path = "../../common/workspace-hack" }
-move-ir-types = { path = "../move-ir/types" }
-diem-framework-releases = { path = "../diem-framework/releases" }
-move-binary-format = { path = "../move-binary-format" }
-move-core-types = { path = "../move-core/types" }
+functional-tests = { path = "../testing-infra/functional-tests" }
 
+datatest-stable = "0.1.1"
 
 [[test]]
 name = "testsuite"
