+++ language/testing-infra/functional-tests/Cargo.toml
@@ -15,12 +15,15 @@ hex = "0.4.3"
 diem-state-view = { path = "../../../storage/state-view" }
 diem-types = { path = "../../../types" }
 diem-vm = { path = "../../diem-vm" }
+move-lang = { path = "../../move-lang" }
 move-binary-format = { path = "../../move-binary-format" }
 bytecode-verifier = { path = "../../bytecode-verifier" }
 language-e2e-tests = { path = "../e2e-tests" }
 diem-config = { path = "../../../config", features = ["fuzzing"] }
 diem-crypto = { path = "../../../crypto/crypto" }
 diem-workspace-hack = { path = "../../../common/workspace-hack" }
+ir-to-bytecode = { path = "../../compiler/ir-to-bytecode" }
+move-ir-types = { path = "../../move-ir/types" }
 once_cell = "1.7.2"
 regex = { version = "1.4.3", default-features = false, features = ["std", "perf"] }
 thiserror = "1.0.24"
@@ -34,4 +37,5 @@ diem-framework-releases = { path = "../../diem-framework/releases" }
 itertools = "0.10.0"
 difference = "2.0.0"
 term_size = "0.3.2"
+tempfile = "3.2.0"
 vm-genesis = { path = "../../tools/vm-genesis" }
