+++ language/testing-infra/functional-tests/src/harness/mvir_harness.rs
@@ -1,0 +1,68 @@
+// Copyright (c) The Diem Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::{
+    compiler::{Compiler, ScriptOrModule},
+    testsuite,
+};
+use anyhow::Result;
+use diem_types::account_address::AccountAddress;
+use ir_to_bytecode::{
+    compiler::{compile_module, compile_script},
+    parser::parse_script_or_module,
+};
+use move_binary_format::CompiledModule;
+use move_core_types::language_storage::ModuleId;
+use move_ir_types::ast;
+use std::{collections::HashMap, path::Path};
+
+struct IRCompiler {
+    deps: HashMap<ModuleId, CompiledModule>,
+}
+
+impl IRCompiler {
+    fn new(diem_framework_modules: Vec<CompiledModule>) -> Self {
+        let deps = diem_framework_modules
+            .into_iter()
+            .map(|m| (m.self_id(), m))
+            .collect();
+        IRCompiler { deps }
+    }
+}
+
+impl Compiler for IRCompiler {
+    /// Compile a transaction script or module.
+    fn compile<Logger: FnMut(String)>(
+        &mut self,
+        mut log: Logger,
+        address: AccountAddress,
+        input: &str,
+    ) -> Result<ScriptOrModule> {
+        Ok(match parse_script_or_module("unused_file_name", input)? {
+            ast::ScriptOrModule::Script(parsed_script) => {
+                log(format!("{}", &parsed_script));
+                ScriptOrModule::Script(
+                    None,
+                    compile_script(Some(address), parsed_script, self.deps.values())?.0,
+                )
+            }
+            ast::ScriptOrModule::Module(parsed_module) => {
+                log(format!("{}", &parsed_module));
+                let module = compile_module(address, parsed_module, self.deps.values())?.0;
+                self.deps.insert(module.self_id(), module.clone());
+                ScriptOrModule::Module(module)
+            }
+        })
+    }
+
+    fn use_compiled_genesis(&self) -> bool {
+        true
+    }
+}
+
+pub fn run_mvir_functional_test(path: &Path) -> datatest_stable::Result<()> {
+    testsuite::functional_tests(
+        IRCompiler::new(diem_framework_releases::current_modules().to_vec()),
+        path,
+    )
+}
